fix: tratamento de excecao para categoria com pratos associados

+++ src/main/java/com/yourmenu/yourmenu_api/category/services/CategoryService.java
@@ -2,18 +2,18 @@ package com.yourmenu.yourmenu_api.category.services;
 
 import com.yourmenu.yourmenu_api.category.Category;
 import com.yourmenu.yourmenu_api.category.CategoryRepository;
+import com.yourmenu.yourmenu_api.category.Exceptions.CategoryAssociatedWithDishException;
 import com.yourmenu.yourmenu_api.category.Exceptions.CategoryNotFoundException;
 import com.yourmenu.yourmenu_api.category.dto.CategoryDTO;
 import com.yourmenu.yourmenu_api.category.dto.CategorySaveDTO;
 import com.yourmenu.yourmenu_api.category.mapper.CategoryMapper;
 import com.yourmenu.yourmenu_api.category.validation.CategoryValidateService;
+import com.yourmenu.yourmenu_api.dish_sizeOptions.dish.DishRepository;
 import com.yourmenu.yourmenu_api.restaurant.Restaurant;
 import com.yourmenu.yourmenu_api.restaurant.RestaurantRepository;
-import com.yourmenu.yourmenu_api.restaurant.RestaurantService;
 import com.yourmenu.yourmenu_api.restaurant.RestaurantValidateService;
 import com.yourmenu.yourmenu_api.restaurant.exception.RestaurantNotFoundException;
 import jakarta.transaction.Transactional;
-import org.apache.commons.lang3.RuntimeEnvironment;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -33,9 +33,12 @@ public class CategoryService {
     @Autowired
     private RestaurantValidateService restaurantValidateService;
 
+    @Autowired
+    private DishRepository dishRepository;
+
     @Transactional
     public CategoryDTO save(CategorySaveDTO dto, String restaurantId, String adminId){
-        Restaurant restaurant = restaurantRepository.findById(restaurantId).orElseThrow(()-> new RestaurantNotFoundException());
+        Restaurant restaurant = restaurantRepository.findById(restaurantId).orElseThrow(RestaurantNotFoundException::new);
         Category category = CategoryMapper.toEntity(dto, restaurant);
         categoryValidateService.validateAdminCanSaveCategory(category, adminId);
         categoryRepository.save(category);
@@ -46,7 +49,7 @@ public class CategoryService {
     public CategoryDTO update(CategorySaveDTO dto, String restaurantId, Long categoryId, String adminId) {
         categoryValidateService.validateAdminCanEditCategory(categoryId, adminId);
         Category oldCategory = categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new CategoryNotFoundException());
+                .orElseThrow(CategoryNotFoundException::new);
         //no futuro deverá ser crida uma função para essa parte de criação de  categoria
         Category newCategory = CategoryMapper.copyEntity(oldCategory);
         newCategory.setName(dto.name().toLowerCase());
@@ -60,7 +63,7 @@ public class CategoryService {
         categoryValidateService.validateCategoryExists(categoryId);
         categoryValidateService.validateCategorybelongsToRestaurant(categoryId, restaurantId);
         Category category = categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new CategoryNotFoundException());
+                .orElseThrow(CategoryNotFoundException::new);
         return CategoryMapper.toDto(category);
     }
 
@@ -74,12 +77,19 @@ public class CategoryService {
     @Transactional
     public CategoryDTO delete(Long categoryId, String adminId) {
         categoryValidateService.validateAdminCanEditCategory(categoryId, adminId);
-        Category category = categoryRepository.findById(categoryId)
+        Category category = categoryRepository
+                .findById(categoryId)
                 .orElseThrow(CategoryNotFoundException::new);
+
+        boolean hasDishes = dishRepository.existsByCategoryId(categoryId);
+        if (hasDishes) {
+            throw new CategoryAssociatedWithDishException();
+        }
         categoryRepository.delete(category);
         return CategoryMapper.toDto(category);
     }
 
+    @Transactional
     public void createCategoriesDefault(Restaurant restaurant) {
         categoryRepository.save(new Category(null, restaurant, "Marmitas"));
         categoryRepository.save(new Category(null, restaurant, "Hamburgueres"));
